Refactor components into component module groups
