+++ flanders/components/content_view.py
@@ -1,0 +1,28 @@
+import discord
+
+
+class TVContentView(discord.ui.LayoutView):
+    def __init__(self, content_url: str, episode_title: str, episode_url: str, episode_details: str, author: str):
+        super().__init__()
+        self.content_url = content_url
+
+        container = discord.ui.Container()
+
+        # Episode title (Simpsons Roasting on an Open Fire)
+        title = discord.ui.TextDisplay(content=f"### **[{episode_title}]({episode_url})**")
+        container.add_item(title)
+
+        # Episode details (S01E01 - 00:00)
+        details = discord.ui.TextDisplay(f"-# {episode_details}")
+        container.add_item(details)
+
+        # Content gif/comic
+        tv_content = discord.ui.MediaGallery()
+        tv_content.add_item(media=self.content_url)
+        container.add_item(tv_content)
+
+        # Generated by @username
+        footer = discord.ui.TextDisplay(content=f"-# *Generated by {author}*")
+        container.add_item(footer)
+
+        self.add_item(container)
